Update Symfony version and add new packages
This commit updates the version of Symfony to 5.4 in the composer.json file and adds new packages, including doctrine and phpdocumentor. Changes are also made to the versions of some packages in the require-dev section. Moreover, modifications are made to the .gitignore, templates/home/index.html.twig, and symfony.lock files.